Refactor agents with an abstract class that enables the creation of different types of agents
